% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConditionalShannonEntropy}
\alias{ConditionalShannonEntropy}
\title{Conditional Shannon Entropy}
\usage{
ConditionalShannonEntropy(condition, feature)
}
\arguments{
\item{condition}{an vector, the condition}

\item{feature}{an vector, the interested feature}
}
\value{
a scalar
}
\description{
Calculate Conditional Shannon Entropy, weighted by its proportion
on the condition
}
\examples{
set.seed(926)
condition <- sample(1:100, 50)
feature <- sample(1:100, 50)
ConditionalShannonEntropy(condition,feature)
}
