% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{DownSamplePrep}
\alias{DownSamplePrep}
\title{Prepare a Downsample}
\usage{
DownSamplePrep(seu, n_clusters, condition = NA, root_cells = NA, ...)
}
\description{
Prepare a downsample for the comparison
}
