% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtractMTX}
\alias{ExtractMTX}
\title{Extract metadata from mtx.gz file}
\usage{
ExtractMTX(file_path)
}
\arguments{
\item{file_path}{File path to \code{matrix.mtx.gz}}
}
\value{
A list of metadata
}
\description{
Extract metadata from mtx.gz file
}
