% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastF.R
\name{fastF_bam2db}
\alias{fastF_bam2db}
\title{Convert BAM to SQLite UMI Matrix}
\usage{
fastF_bam2db(
  bam,
  feature,
  barcode,
  out = ".",
  dbname = ":memory:",
  cell = 1,
  depth = 1,
  seed = 926L,
  umicopies = FALSE
)
}
\arguments{
\item{bam}{Path to BAM file (required)}

\item{feature}{Path to feature list file (required)}

\item{barcode}{Path to barcode list file (required)}

\item{out}{Output directory (default: ".")}

\item{dbname}{Database name (default: ":memory:")}

\item{cell}{Rate of cell barcode (default: 1.0)}

\item{depth}{Rate of depth (default: 1.0)}

\item{seed}{Random seed (default: 926)}

\item{umicopies}{Whether to store umi.tsv.gz (default: FALSE)}
}
\value{
Invisible NULL. Creates SQLite DB in output directory.
}
\description{
Convert BAM to SQLite UMI Matrix
}
