% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastF.R
\name{fastF_filter}
\alias{fastF_filter}
\title{Filter FastQ Files Using Whitelist}
\usage{
fastF_filter(
  R1,
  R2 = NULL,
  I1 = NULL,
  out = ".",
  whitelist = NULL,
  len = 16L,
  seed = 926L,
  rate = 0,
  allcells = FALSE
)
}
\arguments{
\item{R1}{Path to R1 fastq files (required)}

\item{R2}{Path to R2 fastq files (optional)}

\item{I1}{Path to I1 fastq files (optional)}

\item{out}{Output directory (default: ".")}

\item{whitelist}{Path to whitelist file (required unless allcells=TRUE)}

\item{len}{Length of cell barcode (default: 16)}

\item{seed}{Random seed (default: 926)}

\item{rate}{Rate of reads to keep (default: 0)}

\item{allcells}{Keep all reads with cell barcode (default: FALSE)}
}
\value{
Invisible NULL. Creates I1.fastq.gz, R1.fastq.gz, R2.fastq.gz in output.
}
\description{
Filter FastQ Files Using Whitelist
}
