% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastF.R
\name{fastF_freq}
\alias{fastF_freq}
\title{Find Cell Barcode Whitelist from FastQ}
\usage{
fastF_freq(R1, out = ".", len = 16L, umi = 10L)
}
\arguments{
\item{R1}{Path to R1 fastq files (required)}

\item{out}{Output directory for whitelist (default: ".")}

\item{len}{Length of cell barcode (default: 16)}

\item{umi}{Length of UMI (default: 10)}
}
\value{
Invisible NULL. Creates whitelist.txt in output directory.
}
\description{
Find Cell Barcode Whitelist from FastQ
}
