% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{FastQDesign}
\alias{FastQDesign}
\title{FastQDesign}
\usage{
FastQDesign(
  df_power,
  budget,
  power_threshold,
  flow_capacities,
  flow_costs,
  library_costs
)
}
\arguments{
\item{df_power}{A data frame with power information,
\code{N}(filtered cell number), the cell numbers after filtration in the data analysis(\code{Seurat}) pipeline
\code{R}(reads required to get per filtered cell), the total number of FastQ reads divide by the filtered cell numbers
        the total number may use \code{fastF} to get for each sample
\code{power}(0-1) defined by the weighted average of the Adjusted Rand Index on matched cluster membership,
      the Jaccard Index on cluster DE genes and condition DE genes,
      the correlation index on the matched pseudotime(optional),
      the correlation index on the matched 3d cell embedding(optional)}

\item{budget}{A numeric of budget}

\item{power_threshold}{A numeric of power threshold}

\item{flow_capacities}{A vector of flow capacities}

\item{flow_costs}{A vector of flow costs}

\item{library_costs}{A vector of library costs}
}
\value{
A list with power information under constraints and ggplots
}
\description{
Design single-cell RNA experiment with raw FastQ reads
}
