% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degene.R
\name{FindAllMarkersByCondition}
\alias{FindAllMarkersByCondition}
\title{Find Markers By Condition}
\usage{
FindAllMarkersByCondition(seu, condition, ...)
}
\arguments{
\item{seu}{A Seurat object}

\item{condition}{A character name in the `meta.data` of `seu`}

\item{...}{The argument could pass to \code{\link{FindMarkers}}}
}
\value{
A data frame includes DE markers
}
\description{
Find Differential Expressed(DE) markers by Condition
}
