% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{FixedNumClusters}
\alias{FixedNumClusters}
\title{Customize Cluster Number}
\usage{
FixedNumClusters(seu, n_clusters, verbose = FALSE)
}
\arguments{
\item{seu}{a Seurat object}

\item{n_clusters}{scalar, the desired number of clusters}

\item{verbose}{default is \code{\link{FALSE}}}
}
\description{
Customize the number of cluster in the Seurat object
}
