% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JaccardIndex}
\alias{JaccardIndex}
\title{Jaccard Index}
\usage{
JaccardIndex(a, b)
}
\arguments{
\item{a}{an vector or list}

\item{b}{an vector or list}
}
\value{
scalar, jaccard index
}
\description{
Calculate the Jaccard index between two sets
}
\examples{
set.seed(926)
a <- sample(1:100, 50)
b <- sample(1:100, 50)
JaccardIndex(a,b)
}
