% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degene.R
\name{MarkerGeneFilter}
\alias{MarkerGeneFilter}
\title{DE Markers Filters}
\usage{
MarkerGeneFilter(
  df_MarkerGene,
  pct_1 = 0.2,
  pct_2 = 0.2,
  p_val_adj_ = 0.05,
  avg_log2FC_abs = 0
)
}
\arguments{
\item{df_MarkerGene}{a `data.fram` generated from `FindMarkers`}

\item{pct_1}{A minimum cutoff for `pct.1`}

\item{pct_2}{A minimum cutoff for `pct.2`}

\item{p_val_adj_}{A maximum cutoff for `p_val_adj`}

\item{avg_log2FC_abs}{A minimum cutoff for `abs(avg_log2FC)`}
}
\value{
An vector of genes that pass through the filters
}
\description{
Filter the maker genes with series of filters
}
