% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{OptimalNumClusters}
\alias{OptimalNumClusters}
\title{Find the optimal number of cell clusters for Seurat object}
\usage{
OptimalNumClusters(seu, reduction = c("pca", "umap"))
}
\arguments{
\item{seu}{A seurat objet}

\item{reduction}{Reduction to use as input for \code{fpc:prediction.strength}}
}
\value{
An object of class \code{predstr}, please see \code{fpc} for the detail
}
\description{
Find the optimal number of cell clusters using \code{fpc::prediction.strenth}
}
