% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PPV}
\alias{PPV}
\title{Positive Predictive Value(PPV)}
\usage{
PPV(a, b)
}
\arguments{
\item{a}{an vector or list, predicted value}

\item{b}{an vector or list, true value}
}
\value{
scalar, PPV index
}
\description{
Calculate the PPV index between two sets
}
\examples{
set.seed(926)
a <- sample(1:100, 50)
b <- sample(1:100, 50)
PPV(a,b)
}
