% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{RootNodeSelect}
\alias{RootNodeSelect}
\title{Root Node Selection for \code{cell_data_set}}
\usage{
RootNodeSelect(cds, root_cells_ref = NA)
}
\arguments{
\item{cds}{A \code{cell_data_set} object from \code{monocle3}}

\item{root_cells_ref}{An vector of cell ids can be used as the root of cell trajectory}
}
\value{
A character, rootnode id
}
\description{
Select the root node for \code{cell_data_set} based on the given \code{root_cells_ref}
}
