% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{SampleMatch}
\alias{SampleMatch}
\title{Match Downsample to Reference}
\usage{
SampleMatch(reference_list, downsample_list, ...)
}
\description{
Match the features of downsample to the reference, cluster membership,
differential expressed genes by cluster and condition, pseudotime,
cell lower dimension embedding
}
