% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastF.R
\docType{package}
\name{FastQDesign-package}
\alias{FastQDesign-package}
\title{FastQDesign: scRNA-seq Experiment Design Tool}
\description{
Guide investigators in designing single-cell RNA sequencing experiments by shifting focus from UMI matrix back to raw FastQ reads. Includes the fastF C tool for pseudo-design dataset generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yuw444/FastQDesign}
  \item \url{https://yuw444.github.io/FastQDesign/}
  \item Report bugs at \url{https://github.com/yuw444/FastQDesign/issues}
}

}
\author{
\strong{Maintainer}: Yu Wang \email{yuw444@gmail.com}

}
\keyword{internal}
