% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{FastQDesign}
\alias{FastQDesign}
\title{FastQDesign}
\usage{
FastQDesign(
  df_power,
  budget,
  power_threshold,
  reads_valid_rate,
  flowcell_capacities,
  flow_costs,
  library_costs,
  cell_increment = NA,
  read_increment = NA
)
}
\arguments{
\item{df_power}{A data frame with power information,
\code{N}(filtered cell number), the cell numbers after filtration in the data analysis(\code{Seurat}) pipeline
\code{R}(reads required to get per filtered cell), the total number of FastQ reads divide by the filtered cell numbers
        the total number may use \code{fastF} to get for each sample
\code{power}(0-1) defined by the weighted average of the Adjusted Rand Index on matched cluster membership,
      the Jaccard Index on cluster DE genes and condition DE genes,
      the correlation index on the matched pseudotime(optional),
      the correlation index on the matched 3d cell embedding(optional)}

\item{budget}{A numeric of budget}

\item{power_threshold}{A numeric of power threshold}

\item{reads_valid_rate}{The percentage of FastQ reads that is valid when converted to UMIs in the FastQ reference
\code{./filtered_feature_bc_matrix/barcodes.tsv.gz} after the alignment}

\item{flow_costs}{A vector of flow costs}

\item{library_costs}{A vector of library costs}

\item{cell_increment}{A numeric of cell increment for the design, default is \code{floor(max(df_power$N) / 50) * 5}}

\item{read_increment}{A numeric of read increment for the design, default is \code{floor(max(df_power$R) / 100) * 10}}

\item{flow_capacities}{A vector of flow capacities}
}
\value{
A list with power information under constraints and ggplots
}
\description{
Design single-cell RNA experiment with raw FastQ reads
}
