% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{FindPseudotime}
\alias{FindPseudotime}
\title{Find Pseudotime for \code{Seurat} Object}
\usage{
FindPseudotime(
  seu,
  interactive = FALSE,
  use_partition = FALSE,
  root_cells_ref = NA,
  min_branch_len = 10,
  redo_sctransform = FALSE,
  vars_to_regress = NULL
)
}
\arguments{
\item{seu}{A \code{Seurat} object}

\item{interactive}{Whether to choose the root node interactively}

\item{use_partition}{Whether to \code{use_partition} when \code{learn_graph}}

\item{root_cells_ref}{An vector of cell ids can be used as the root of cell trajectory}

\item{min_branch_len}{The minimum branch length in the cell trajectory, the same parameter in \code{monocle3::learn_graph}}

\item{redo_sctransform}{Whether to redo the \code{Seurat::SCTransform}}

\item{vars_to_regress}{A character string, used in \code{Seurat::SCTransform}}
}
\value{
A \code{cell_data_set} object with pseudotime as feature
}
\description{
Find the pseudotime of a \code{Seurat} object based on its \code{UMAP}.
}
