% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{SamplePrep}
\alias{SamplePrep}
\title{Prepare a sample}
\usage{
SamplePrep(
  seu,
  cluster_id = NA,
  n_clusters = NA,
  use_default_res = TRUE,
  condition = NA,
  cell_3d_embedding = FALSE,
  pca_used = 1:30,
  vars_to_regress = NULL,
  use_partition = FALSE,
  close_loop = FALSE,
  learn_graph_control = NULL,
  interactive = FALSE,
  root_cells_ref = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{seu}{A \code{Seurat} object}

\item{cluster_id}{A A character, the condition colname in \code{Seurat@meta.data}}

\item{n_clusters}{A scalar, the desired number of clusters}

\item{use_default_res}{Whether to use the default resolution in \code{Seurat::FindClusters}}

\item{condition}{A character, the condition colname in \code{Seurat@meta.data}}

\item{cell_3d_embedding}{Calculate 3d embedding for the input \code{seu}}

\item{pca_used}{Used in \code{Seurat::FindNeighbors} and \code{Seurat::RunUMAP}}

\item{vars_to_regress}{A character string, used in \code{Seurat::SCTransform}}

\item{use_partition}{Whether to \code{use_partition} when \code{monocle3::learn_graph}}

\item{close_loop}{Logic, a parameter in \code{monocle3::learn_graph}}

\item{learn_graph_control}{List, a parameter in \code{monocle3::learn_graph}}

\item{interactive}{Logic, whether to select root cell interactively when considering cell fate}

\item{root_cells_ref}{An vector of cell ids can be used as the root of cell trajectory}
}
\value{
list Seurat with extra meta.data features,
 data.frames of DE genes by cluster and condition
}
\description{
Prepare a sample for the comparison
}
