% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{SamplePrep}
\alias{SamplePrep}
\title{Prepare a sample}
\usage{
SamplePrep(
  seu,
  n_clusters = NA,
  condition = NA,
  cell_3d_embedding = FALSE,
  pca_used = 1:30,
  interactive = FALSE,
  root_cells_ref = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{seu}{A \code{Seurat} object}

\item{n_clusters}{A scalar, the desired number of clusters}

\item{condition}{A character, the condition colname in \code{Seurat@meta.data}}

\item{root_cells_ref}{An vector of cell ids can be used as the root of cell trajectory}
}
\value{
list Seurat with extra meta.data features,
 data.frames of DE genes by cluster and condition
}
\description{
Prepare a sample for the comparison
}
